requirements.txt
# Django
Django
django-environ
psycopg2-binary

# REST Framework
djangorestframework
drf-spectacular

# Authentication
django-allauth
dj-rest-auth
<<<<<<< HEAD
django-rest-framework-simplejwt
=======
djangorestframework-simplejwt
>>>>>>> f3f460e4d9735213c1a8a8cc1b9cec37ca680d72

# Database
django-postgres-extra

# CORS
django-cors-headers

# Task queue
celery
redis
django-celery-results
django-celery-beat

# Email
django-anymail

# Development
black
flake8
isort

# Production
gunicorn
whitenoise
<<<<<<< HEAD
=======
huggingface-hub==0.19.4
>>>>>>> f3f460e4d9735213c1a8a8cc1b9cec37ca680d72
